FastaProcessing

Documentation for FastaProcessing.

FastaProcessing.combined_dna_dataset

FastaProcessing.multi_2_multi_dna_dataset

FastaProcessing.multi_2_one_dna_dataset

FastaProcessing.get_all_reads

FastaProcessing.get_range_dataset

FastaProcessing.obtain_dna_heads_and_dna_reads

FastaProcessing.trim_dna_reads!

FastaProcessing.combined_dna_dataset — Type

dnadatasets: Vector of dnadata read_range: e.g. dataset A has 500 reads, dataset B has 1000 reads, get Dict(1=>1:500, 2=>501:1500) L: length of the sequence for all the datasets (yes, they must have the same length)

combined the dna reads from all datasets following the order of the input dnadatasets useful for 1. referencing the different crosslinking signitures (readrange) 2. generate a datamatrix that includes all of the sequences

FastaProcessing.multi_2_multi_dna_dataset — Type

multi2multidnadataset multiple datasets of dna seuqneces, e.g. dataset1: [s₁, s₂,...,sₙ] dataset2: [t₁, t₂,...,tₘ] and multiple label sets labels1: [l₁, l₂,...,lₙ] labels2: [l₁, l₂,...,lₘ]

note that each dataset and its corresponding label set 
must have the same cardinality

FastaProcessing.multi_2_one_dna_dataset — Type

multi2onednadataset multiple datasets of dna seuqneces, e.g. dataset1: [s₁, s₂,...,sₙ] dataset2: [t₁, t₂,...,tₙ] and one label set labels: [l₁, l₂,...,lₙ]

note that all datasets and the label set must have the same cardinality

This can happen in splicing datasets, where we only want to 
model multiple (non-endogeneous) intronic regions and how 
they affect gene-expression

FastaProcessing.get_all_reads — Method

get_all_reads

FastaProcessing.get_range_dataset — Method

example: dataset A has 500 reads dataset B has 1000 reads return Dict(1=>1:500, 2=>501:1500)

FastaProcessing.obtain_dna_heads_and_dna_reads — Method

obtaindnaheadsanddna_reads

return:
    dna_heads: header of all the fasta sequences (Vector)
    dna_reads: dna strings

FastaProcessing.trim_dna_reads! — Method

trimdnareads! trim the dna reads on both sides by n bases; get rid of reads that contain n or N in the shortened read

return:
    dna_heads: header of all the fasta sequences (Vector) that are kept
    dna_reads: trimmed dna sequences
    len: len of reads after trimming